% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{convertPatchCoordsToSpatialPoints}
\alias{convertPatchCoordsToSpatialPoints}
\title{convert coordinates of deep feature patches to image-based physical coordinates}
\usage{
convertPatchCoordsToSpatialPoints(patchCoords, img, patchSize = 32)
}
\arguments{
\item{patchCoords}{input patch coordinates (usually from \code{deepFeatures})}

\item{img}{reference image defining spatial domain}

\item{patchSize}{vector or scalar defining patch dimensions}
}
\value{
spatial patch coordinates
}
\description{
High-level function for handling the converion between voxel and corner-based
indices of patches to center and physical-based landmark coordinates.
}
\examples{

library(ANTsR)
img <- ri( 1 ) \%>\% iMath( "Normalize" )
mask = randomMask( getMask( img ), 20 )
features = deepFeatures( img, mask, patchSize = 32 )
coords = convertPatchCoordsToSpatialPoints( features$patchCoords, img )

}
\author{
Avants BB
}
