% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{coordinateImages}
\alias{coordinateImages}
\title{make a coordinate image}
\usage{
coordinateImages(mask, physicalCoordinates = TRUE)
}
\arguments{
\item{mask}{defining wherein we create coordinates}

\item{physicalCoordinates}{select index or phyiscal coordinates}
}
\value{
list of coordinate images
}
\description{
This function will return n-dimensionality coordinate images that encode
the spatial ITK coordinates of the image domain.
}
\examples{

library(ANTsR)
img <- ri( 1 )
coords = coordinateImages( img * 0 + 1, TRUE )

}
\author{
Avants BB
}
