% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{deepLandmarkRegressionWithHeatmaps}
\alias{deepLandmarkRegressionWithHeatmaps}
\title{Deep heatmap-based landmark regression stage}
\usage{
deepLandmarkRegressionWithHeatmaps(
  model,
  activation = c("none", "relu", "trelu", "softmax", "sigmoid"),
  theta
)
}
\arguments{
\item{model}{input deep model, presumably a unet.  it should have an input
dimensionality that is dimension-appropriate.  the number of output channels
should match the number of landmarks and should be an image.}

\item{activation}{the activation function for the regression maps}

\item{theta}{the theta parameter for thresholded relu}
}
\value{
the augmented model
}
\description{
Deep heatmap-based landmark regression stage
}
