% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{deepPatchMatchMultiStart}
\alias{deepPatchMatchMultiStart}
\title{deepPatchMatch with multiple starting points}
\usage{
deepPatchMatchMultiStart(
  movingImage,
  fixedImage,
  movingImageMask,
  fixedImageMask,
  movingPatchSize = 16,
  fixedPatchSize = 16,
  block_name = 20,
  vggmodel,
  subtractor = 127.5,
  patchVarEx = 0.95,
  meanCenter = FALSE,
  numberOfStarts = 1,
  sdAffine = 5,
  verbose = FALSE
)
}
\arguments{
\item{movingImage}{input image from which we extract patches that are
transformed to the space of the fixed image}

\item{fixedImage}{input image that provides the fixed reference domain.}

\item{movingImageMask}{defines the object of interest in the movingImage}

\item{fixedImageMask}{defines the object of interest in the fixedImage}

\item{movingPatchSize}{integer greater than or equal to 32.}

\item{fixedPatchSize}{integer greater than or equal to 32.}

\item{block_name}{name of vgg feature block, either block2_conv2 or integer.
use the former for smaller patch sizes.}

\item{vggmodel}{prebuilt feature model}

\item{subtractor}{value to subtract when scaling image intensity; should be
chosen to match training paradigm eg 127.5 for vgg and 0.5 for resnet like.}

\item{patchVarEx}{patch variance explained for ripmmarc}

\item{meanCenter}{boolean to mean center each patch for ripmmarc}

\item{numberOfStarts}{the number of starting points to try}

\item{sdAffine}{standard deviation parameter e.g. 0.15}

\item{verbose}{boolean}
}
\value{
matched points and transformation
}
\description{
High-level function for deep patch matching that uses multiple starting
points to reduce sensitivity to initialization.
}
\examples{
\dontrun{
library( tensorflow )
library( ANTsR )
nP1 = 250
nP2 = 500
psz = 12
img <- ri( 1 ) \%>\% iMath( "Normalize" )
img2 <- ri( 2 ) \%>\% iMath( "Normalize" )
mask = randomMask( getMask( img ), nP1 )
mask2 = randomMask( getMask( img2 ), nP2 )
match = deepPatchMatchMultiStart( img2, img, mask, mask2, numberOfStarts=3, verbose=T )
}
}
\author{
Avants BB
}
