% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{featureDistanceMap}
\alias{featureDistanceMap}
\title{Feature distance map}
\usage{
featureDistanceMap(image1, image2, jointMask, patchSize = 32, ...)
}
\arguments{
\item{image1}{image 1}

\item{image2}{image 2}

\item{jointMask}{mask covering features of both images}

\item{patchSize}{patchSize for both images}

\item{...}{additional options to pass to deepFeatures e.g. vggmodel, block_name, substractor}
}
\value{
error map
}
\description{
Feature distance map
}
