% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{fitTransformToPairedPoints}
\alias{fitTransformToPairedPoints}
\title{Fit transform to points}
\usage{
fitTransformToPairedPoints(movingPoints, fixedPoints,
  transformType = "Affine", lambda = 1e-04)
}
\arguments{
\item{movingPoints}{moving points matrix}

\item{fixedPoints}{fixed points matrix}

\item{transformType}{Affine, Rigid and Similarity currently supported}

\item{lambda}{ridge penalty}
}
\value{
antsTransform that maps the moving image to the fixed image space.
the inverse transform maps the moving points to the fixed space. associated
error is also returned.'
}
\description{
This function will use either the Kabsch algorithm or a least squares fitting
algorithm to match the pairs of points that the user provides.  An antsr
transform is returned.
}
