% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{fitTransformToPairedPointsTF}
\alias{fitTransformToPairedPointsTF}
\title{Fit transform to points with tensorflow}
\usage{
fitTransformToPairedPointsTF(
  movingPoints,
  fixedPoints,
  numberOfPoints,
  dimensionality,
  transformType = c("Rigid", "Similarity", "Affine"),
  batch_size = 1,
  preventReflection = TRUE
)
}
\arguments{
\item{movingPoints}{moving points matrix}

\item{fixedPoints}{fixed points matrix}

\item{numberOfPoints}{per sample in batch}

\item{dimensionality}{of the point set}

\item{transformType}{Rigid, Similarity or Affine currently supported}

\item{batch_size}{the batch size}

\item{preventReflection}{boolean}
}
\value{
tensorflow tensor objects containing the transformation matrices
as the first entry in the list and the error term as the second.
}
\description{
This function will use either the Kabsch algorithm or a least squares fitting
algorithm to match the pairs of points that the user provides.  A tensorflow
tensor is returned.
}
