% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{matchedLandmarks}
\alias{matchedLandmarks}
\title{Convert match output to landmarks in physical space}
\usage{
matchedLandmarks(matchObject, referenceImage, movingImage, patchSize)
}
\arguments{
\item{referenceImage}{the fixed image}

\item{movingImage}{the image that will be matched to the fixed image}

\item{match}{object, the output of \code{deepPatchMatch}}

\item{patch}{size of patch features}
}
\value{
output list contains fixed and matched points
}
\description{
Convert match output to landmarks in physical space
}
