% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{matchedPatches}
\alias{matchedPatches}
\title{extract matched patches between two images}
\usage{
matchedPatches(movingImage, fixedImage, patchMatchOutput,
  fixedPatchRadius = 31, verbose = FALSE)
}
\arguments{
\item{movingImage}{input image from which we extract patches that are
transformed to the space of the fixed image}

\item{fixedImage}{input image that provides the fixed reference domain.}

\item{patchMatchOutput}{the data frame output from \code{patchMatch}.}

\item{fixedPatchRadius}{integer greater than zero.}

\item{verbose}{boolean, will print to screen.}
}
\value{
lists of corresponding patches
}
\description{
provides the matched patches given output of \code{patchMatch}
}
\examples{

library(ANTsR)
img <- ri( 1 ) \%>\% iMath( "Normalize" )
img2 <- ri( 2 ) \%>\% iMath( "Normalize" )
mask = randomMask( getMask( img ), 2 )
match = patchMatch( img2, img, mask, fixedPatchRadius = 3 )
myMatches = matchedPatches( img2, img, match, fixedPatchRadius = 3 )

}
\author{
Avants BB
}
