% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{patchMatch}
\alias{patchMatch}
\title{patch match two images}
\usage{
patchMatch(movingImage, fixedImage, fixedImageMask,
  finalTransform = "Rigid", fixedPatchRadius = 31, initialMap,
  visualize = FALSE, verbose = FALSE)
}
\arguments{
\item{movingImage}{input image from which we extract patches that are
transformed to the space of the fixed image}

\item{fixedImage}{input image that provides the fixed reference domain.}

\item{fixedImageMask}{defines the object of interest in the fixedImage}

\item{finalTransform}{defaults to "Rigid" but can be any \code{antsRegistration}.}

\item{fixedPatchRadius}{integer greater than zero.}

\item{initialMap}{this is the output of a call to \code{antsRegistration}.
if it is not present, a quick SyN map will be computed.}

\item{visualize}{boolean, will plot to screen}

\item{verbose}{boolean, will print to screen}
}
\value{
data frame of corresponding points
}
\description{
High-level function for patch matching that makes many assumptions and
therefore minimizes the number of parameters the user needs to choose.
This prioritizes usability at the cost of optimality.
}
\examples{

library(ANTsR)
img <- ri( 1 ) \%>\% iMath( "Normalize" )
img2 <- ri( 2 ) \%>\% iMath( "Normalize" )
mask = randomMask( getMask( img ), 2 )
match = patchMatch( img2, img, mask, fixedPatchRadius = 3 )

}
\author{
Avants BB
}
