% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{polarDecomposition}
\alias{polarDecomposition}
\title{polar decomposition with reflection control}
\usage{
polarDecomposition(X)
}
\arguments{
\item{X}{matrix}
}
\value{
decomposition into P Z and approximation of X by \code{P \%*\% Z}
}
\description{
decomposes X into P and Z (matrices) where Z is a rotation and P is shearing.
will prevent Z from containing reflections.
}
\examples{

pd = polarDecomposition( matrix( rnorm(9), nrow=3 ) )

}
\author{
Avants BB
}
