% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{randomAffineImage}
\alias{randomAffineImage}
\title{transform an image by a random matrix}
\usage{
randomAffineImage(
  image,
  txtype = c("Rigid", "Affine", "ScaleShear"),
  sdAffine = 0.1,
  seeder,
  fixedParams,
  interpolation = "nearestNeighbor"
)
}
\arguments{
\item{image}{input image to be transformed}

\item{txtype}{one of Rigid, Affine and ScaleShear}

\item{sdAffine}{standard deviation parameter e.g. 0.15}

\item{seeder}{random seed}

\item{fixedParams}{optional fixed parameters for ANTs or ITK transformation}

\item{interpolation}{optional interpolation choice}
}
\value{
the transformed image and the transformation matrix
}
\description{
generates a random transformation matrix and its application to an image
}
\examples{
rr = randomAffineImage( ri( 1 ), "Rigid", sdAffine = 5 )
}
\author{
Avants BB
}
