% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{RANSAC}
\alias{RANSAC}
\title{Random sample consensus (RANSAC)}
\usage{
RANSAC(
  fixedPoints,
  movingPoints,
  transformType = "Affine",
  minNtoFit = 16,
  maxIterations = 20,
  errorThreshold = 1,
  goodProportion = 0.5,
  lambda = 1e-08,
  verbose = FALSE
)
}
\arguments{
\item{fixedPoints}{fixed points matrix}

\item{movingPoints}{moving points matrix}

\item{transformType}{Affine, Rigid and Similarity currently supported}

\item{minNtoFit}{the minimum number of data values required to fit the model.}

\item{maxIterations}{the maximum number of iterations allowed in the algorithm}

\item{errorThreshold}{a threshold value for determining when a test data point fits a model.
this parameter is set based on the standard deviation in the random subset model.
that is, a point fits the model error distribution if it is within the bracket
of values between mean error plus or minus sd error times errorThreshold.}

\item{goodProportion}{the fraction of close data values required to assert that a model
fits well to data.  that is, if equal to 0.5, then one would need 50 points to
assert that a model fit is good if the whole dataset contains 100 points.}

\item{lambda}{ridge penalty in zero to one}

\item{verbose}{boolean}
}
\value{
output list contains best fitted model, inliers, outliers
}
\description{
Random sample consensus (RANSAC)
}
