% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{fitTransformToPairedPoints}
\alias{fitTransformToPairedPoints}
\title{Fit transform to points}
\usage{
fitTransformToPairedPoints(movingPoints, fixedPoints,
  transformType = "Affine", lambda = 1e-04)
}
\arguments{
\item{movingPoints}{moving points matrix}

\item{fixedPoints}{fixed points matrix}

\item{transformType}{affine, rigid and similarity currently supported}

\item{lambda}{ridge penalty}
}
\value{
antsTransform
}
\description{
Fit transform to points
}
