% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{fitTransformToPairedPoints}
\alias{fitTransformToPairedPoints}
\title{Fit transform to points}
\usage{
fitTransformToPairedPoints(movingPoints, fixedPoints, transformType)
}
\arguments{
\item{movingPoints}{moving points matrix}

\item{fixedPoints}{fixed points matrix}

\item{transformType}{e.g. rigid or affine}
}
\value{
antsTransform
}
\description{
Fit transform to points
}
