% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{fitTransformToPairedPoints}
\alias{fitTransformToPairedPoints}
\title{Fit transform to points}
\usage{
fitTransformToPairedPoints(movingPoints, fixedPoints, transformType)
}
\value{
antsTransform
}
\description{
Fit transform to points
}
