% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{matchedLandmarks}
\alias{matchedLandmarks}
\title{Convert match output to landmarks in physical space}
\usage{
matchedLandmarks(matchObject, referenceImage, movingImage, patchSize)
}
\arguments{
\item{referenceImage}{the fixed image}

\item{movingImage}{the image that will be matched to the fixed image}

\item{match}{object, the output of \code{deepPatchMatch}}

\item{patch}{size of patch features}
}
\value{
output list contains fixed and matched points
}
\description{
Convert match output to landmarks in physical space
}
\examples{
\dontrun{
mlm = matchedLandmarks( match, img, img2, c(psz,psz) )
ct = 0
mxct = 18
lmImage1 = img * 0
lmImage2 = img2 * 0
for ( k in 1:nrow( mlm$fixedPoints ) ) {
  if ( ct < mxct ) {
    pt1i = makePointsImage( matrix(mlm$fixedPoints[k,],ncol=2), img, radius = 2 ) * k
    pt2i = makePointsImage( matrix(mlm$movingPoints[k,],ncol=2), img2, radius = 2 ) * k
    lmImage1 = lmImage1 + pt1i
    lmImage2 = lmImage2 + pt2i
    }
  ct = ct + 1
 }

plot( img, lmImage1 )
plot( img2, lmImage2 )
}
}
