% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{RANSAC}
\alias{RANSAC}
\title{Random sample consensus (RANSAC)}
\usage{
RANSAC(data, y_col, model, model_args, n = 2, k = NA, t = NA,
  d = NA, verbose = FALSE)
}
\arguments{
\item{data}{dataframe with target column and model matrix columns}

\item{y_col}{name of target column}

\item{model}{model function to fit}

\item{model_args}{list, arguments to the model}

\item{n}{the minimum number of data values required to fit the model}

\item{k}{the maximum number of iterations allowed in the algorithm}

\item{t}{a threshold value for determining when a data point fits a model}

\item{d}{the number of close data values required to assert that a model fits well to data}
}
\value{
a list of output contains best fitted model, inliers, outliers
}
\description{
Random sample consensus (RANSAC)
}
