% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMatch.R
\name{RANSACAlt}
\alias{RANSACAlt}
\title{Alternative random sample consensus (RANSACAlt)}
\usage{
RANSACAlt(
  fixedPoints,
  movingPoints,
  transformType = "Affine",
  nToTrim = 2,
  minProportionPoints = 0.5,
  lambda = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{fixedPoints}{fixed points matrix}

\item{movingPoints}{moving points matrix}

\item{transformType}{Affine, Rigid and Similarity currently supported}

\item{nToTrim}{the number of points to throw away at each iteration}

\item{minProportionPoints}{the minimum proportion of points to return}

\item{lambda}{ridge penalty}

\item{verbose}{boolean}
}
\value{
output list contains best fitted model, inliers, outliers
}
\description{
Alternative random sample consensus (RANSACAlt)
}
